refactoring : frame directory 변경
